feat: update column config
